impl: chromakey, refactor builders
