+++ test/resources/clips.[].chromakey.yml
@@ -1,0 +1,1 @@
+{"size":{"background_color":"blue"},"clips":[{"file":"./assets/century-leaf-falling-autumn-maple-leaves-falling-maple-autumn-leaves-falling-autumn-leaves-falling-against-black-background-free-video.mp4","trim.start":"3","chromakey":"black"}]}
