+++ test/resources/preview default clip duration.yml
@@ -1,0 +1,1 @@
+{"size":{"background_color":"blue"},"clips":[{"file":"./assets/1636302951890.jpg"},{"file":"./assets/Pexels Videos 2048452.mp4","layout.height":"50%","layout.x":"center","layout.y":"center","crop.width":"600px","crop.height":"600px"}],"preview":"5"}
